requirements.txt
wheel
setuptools

torch==2.1.0
diffusers
Pillow
typing_extensions
sympy
mpmath
packaging
huggingface_hub
tqdm
flask
requests
urllib3
PyYAML
filelock
numpy
safetensors
transformers
regex
accelerate
gradio
